refactor(src, test): split utils.js into constants.js and utils.js
